Add having handling for numerical columns with non aggregate values
