--- frontend/src/components/AdvancedHavingQueryBuilder/utils/utility.ts
@@ -1,11 +1,0 @@
-export const sortAggregateOptions = (options: string[], selected: string): string[] => {
-  if (options.indexOf(selected) === 0) {
-    return options;
-  }
-  const firstOption = options[0];
-  const selectedOptionIndex = options.indexOf(selected);
-  options[0] = selected;
-  options[selectedOptionIndex] = firstOption;
-
-  return options;
-};
